current location, changing the icon and the name of the app
