+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.nearmeproject"
+        applicationId "com.example.nearme"
         minSdk 21
         targetSdk 31
         versionCode 1
@@ -63,7 +63,7 @@ dependencies {
     implementation 'com.zhihu.android:matisse:0.5.3-beta3'
     implementation 'com.github.bumptech.glide:glide:4.12.0'
     annotationProcessor 'com.github.bumptech.glide:compiler:4.12.0'
-
+    implementation 'com.google.android.gms:play-services-location:18.0.0'
