+++ app/src/androidTest/java/com/example/nearme/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.nearmeproject
+package com.example.nearme
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
