+++ app/src/main/java/com/example/nearme/database/database.kt
@@ -1,0 +1,4 @@
+package com.example.nearme.database
+
+class database {
+}
