+++ app/src/main/java/com/example/nearme/model/identity/identity.kt
@@ -1,0 +1,4 @@
+package com.example.nearme.model.identity
+
+class identity {
+}
