+++ app/src/main/java/com/example/nearme/util/util.kt
@@ -1,0 +1,4 @@
+package com.example.nearme.util
+
+class util {
+}
