+++ app/src/main/java/com/example/nearme/views/identity/identity.kt
@@ -1,0 +1,4 @@
+package com.example.nearme.views.identity
+
+class identity {
+}
