+++ app/src/main/java/com/example/nearme/views/main/ImageViewModel.kt
@@ -1,13 +1,12 @@
-package com.example.nearmeproject.views.main
+package com.example.nearme.views.main
 
 import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.nearmeproject.model.images.ImageModel
-import com.example.nearmeproject.model.images.Photo
-import com.example.nearmeproject.model.images.Photos
-import com.example.nearmeproject.repositories.ApiRepo
+import com.example.nearme.model.images.ImageModel
+import com.example.nearme.model.images.Photo
+import com.example.nearme.repositories.ApiRepo
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import java.lang.Exception
@@ -15,19 +14,18 @@ import java.lang.Exception
 private const val TAG = "ImageViewModel"
 class ImageViewModel:ViewModel() {
 
+
     val apiRepo=ApiRepo.get()
-
-
     val imageLiveData=MutableLiveData<ImageModel>()
     val oneimagelivedata=MutableLiveData<Photo>()
 
 
-    fun callImage(){
+    fun callImage(lat:Double, lon:Double){
 
         viewModelScope.launch (Dispatchers.IO){
             try {
 
-                val response=apiRepo.getImages(26.399250,49.984360)
+                val response=apiRepo.getImages(lat,lon)
                 if (response.isSuccessful){
                     response.body()?.run {
                         Log.d(TAG,response.body().toString())
