--- app/src/main/java/com/example/nearmeproject/database/database.kt
@@ -1,4 +1,0 @@
-package com.example.nearmeproject.database
-
-class database {
-}
