--- app/src/main/java/com/example/nearmeproject/model/identity/identity.kt
@@ -1,4 +1,0 @@
-package com.example.nearmeproject.model.identity
-
-class identity {
-}
