--- app/src/main/java/com/example/nearmeproject/util/util.kt
@@ -1,4 +1,0 @@
-package com.example.nearmeproject.util
-
-class util {
-}
