--- app/src/main/java/com/example/nearmeproject/views/identity/identity.kt
@@ -1,4 +1,0 @@
-package com.example.nearmeproject.views.identity
-
-class identity {
-}
